pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "narrative-subcorpora"
version = "0.1.0"
description = "Construct subcorpora of historical text data around events"
readme = "readme.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=0.9",
    "pyarrow>=14.0",
    "pandas>=2.0",
    "click>=8.0",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
nlp = ["spacy>=3.0"]
embeddings = ["sentence-transformers>=2.0"]
dev = ["pytest>=7.0", "ruff>=0.1"]

[project.scripts]
nsc = "narrative_subcorpora.cli:cli"

[tool.setuptools.packages.find]
include = ["narrative_subcorpora*"]
